Added place photo upload functionality.

+++ backend/app/schemas/note.py
@@ -7,7 +7,7 @@ class CommunityNoteRequest(BaseModel):
     outlets_available: Optional[bool] = None
     parking_available: Optional[bool] = None
     food_available: Optional[bool] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     feel: Optional[List[str]] = None
     comment: Optional[str] = None
     image_url: Optional[str] = None
@@ -18,7 +18,7 @@ class NoteUpdate(BaseModel):
     outlets_available: Optional[bool] = None
     parking_available: Optional[bool] = None
     food_available: Optional[bool] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     feel: Optional[List[str]] = None
     comment: Optional[str] = None
     image_url: Optional[str] = None
